shared-workspace/shared-resources/scripts/requirements.txt
# Requires Python 3.8+
pyyaml>=6.0
